Rename `container` module to `mscmap`
Better reflects what it actually is.
