+++ README.md
@@ -69,9 +69,9 @@ functions of the YubiKey:
 | 🚧️ | `certificate` | [#22] | Certificates for stored keys |
 | 🚧 | `chuid`       | [#23] | Cardholder Unique Identifier (CHUID) |
 | ✅️ | `config`      | [#24] | Support for reading on-key configuration |
-| ⚠️ | `container`   | [#25] | MS Container Map Records |
 | 🚧 | `key`         | [#26] | Crypto key management: list, generate, import |
 | 🚧 | `mgm`         | [#26] | Management Key (MGM) support: set, get, derive |
+| ⚠️ | `mscmap`      | [#25] | MS Container Map Records |
 | ⚠️ | `msroots`     | [#28] | `msroots` file: PKCS#7 formatted certificate store for enterprise trusted roots |
 
 Legend:
